read `power-assert-recorder` from disk
