+++ lib/babel-assertion-visitor.js
@@ -6,34 +6,13 @@ var babelgen = require('babel-generator');
 var define = require('define-properties');
 var toBeCaptured = require('./to-be-captured');
 var toBeSkipped = require('./to-be-skipped');
-var helperCode = '(' + (function () {
-      function PowerAssertRecorder() {
-          this.captured = [];
-      }
-
-      PowerAssertRecorder.prototype._capt = function _capt (value, espath) {
-          this.captured.push({value: value, espath: espath});
-          return value;
-      };
-
-      PowerAssertRecorder.prototype._expr = function _expr (value, args) {
-          return {
-              powerAssertContext: {
-                  value: value,
-                  events: this.captured
-              },
-              source: {
-                  content: args.content,
-                  filepath: args.filepath,
-                  line: args.line,
-                  generator: !!args.generator,
-                  async: !!args.async
-              }
-          };
-      };
-
-      return PowerAssertRecorder;
-  }).toString() + ')()';
+var fs = require('fs');
+var helperCode = '(' +
+  fs.readFileSync(require.resolve('./power-assert-recorder.js'), 'utf8')
+    .split('\n')
+    .slice(2)
+    .join('\n')
+  + ')()';
 
 function BabelAssertionVisitor (babel, matcher, options) {
     this.babel = babel;
